backend/requirements.txt
asgiref==3.8.1
atomicwrites==1.4.1
attrs==23.2.0
Brotli==1.1.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
CodeConvert==3.0.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.8
cssselect2==0.7.0
defusedxml==0.8.0rc2
Django==4.2.13
django-admin==2.0.2
django-cors-headers==3.13.0
django-excel-response2==3.0.6
django-extensions==3.2.3
django-filter==24.2
django-models-ext==1.1.11
django-short-url==1.1.8
django-six==1.0.5
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
excel-base==1.0.4
flake8==6.0.0
flake8-isort==6.0.0
fonttools==4.53.0
furl==2.1.3
gunicorn==20.1.0
html5lib==1.1
idna==3.7
iniconfig==2.0.0
isort==5.13.2
isoweek==1.3.3
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
oauthlib==3.2.2
orderedmultidict==1.0.1
packaging==24.0
Pillow==9.0.0
pluggy==0.13.1
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.10.0
pycparser==2.22
pydyf==0.10.0
pyflakes==3.0.1
PyJWT==2.8.0
pyphen==0.15.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0
requests==2.32.3
requests-oauthlib==2.0.0
screen==1.0.1
shortuuid==1.0.13
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.4
sqlparse==0.5.0
TimeConvert==3.0.13
tinycss2==1.3.0
toml==0.10.2
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
webcolors==1.11.1
webencodings==0.5.1
xlwt==1.3.0
zopfli==0.2.3
